Vers. 2.0.6: Fehlerbehebung bei extremen Rauheiten nach Aguirre-Pe/Fuentes, zusätzliche Konfig-Datei im Prof-Ordner möglich.

+++ _import/trunk/fortran/kalypso-1d/source/wspow.f90
@@ -99,7 +99,7 @@ REAL :: hvst
 REAL :: hrst
 REAL :: psieins, psiorts
 INTEGER :: nprof
-INTEGER :: hgrenz
+REAL :: hgrenz
 INTEGER :: ikenn
 INTEGER :: nblatt
 INTEGER :: nz
